remove new_s3.py and s3_w_boto3.py update base s3 bucket reader with efficient contains and iter
