+++ s3dol/__init__.py
@@ -1,10 +1,4 @@
 """
 s3 (through boto3) with a simple (dict-like or list-like) interface
 """
-from s3dol import new_s3
-from s3dol import s3_store
-from s3dol import s3_w_boto3
-
-from s3dol.new_s3 import *
-from s3dol.s3_store import *
-from s3dol.s3_w_boto3 import *
+from s3dol.base import S3Dol
